refactor(mail-folder): move RemoteFolder to :feature:mail:folder:api
