--- feature/folder/api/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-    id(ThunderbirdPlugins.Library.jvm)
-    alias(libs.plugins.android.lint)
-}
-
-dependencies {
-    implementation(projects.feature.mail.folder.api)
-}
