+++ legacy/core/src/main/java/com/fsck/k9/mailstore/SpecialFolderSelectionStrategy.kt
@@ -1,7 +1,7 @@
 package com.fsck.k9.mailstore
 
-import net.thunderbird.feature.folder.api.RemoteFolder
 import net.thunderbird.feature.mail.folder.api.FolderType
+import net.thunderbird.feature.mail.folder.api.RemoteFolder
 
 /**
  * Implements the automatic special folder selection strategy.
