+++ legacy/mailstore/build.gradle.kts
@@ -13,7 +13,6 @@ dependencies {
 
     implementation(projects.feature.mail.account.api)
     implementation(projects.feature.mail.folder.api)
-    implementation(projects.feature.folder.api)
     implementation(projects.feature.search)
 
     implementation(projects.mail.common)
