+++ legacy/ui/folder/src/main/java/app/k9mail/legacy/ui/folder/FolderNameFormatter.kt
@@ -1,9 +1,9 @@
 package app.k9mail.legacy.ui.folder
 
 import android.content.res.Resources
-import net.thunderbird.feature.folder.api.RemoteFolder
 import net.thunderbird.feature.mail.folder.api.Folder
 import net.thunderbird.feature.mail.folder.api.FolderType
+import net.thunderbird.feature.mail.folder.api.RemoteFolder
 
 class FolderNameFormatter(private val resources: Resources) {
     fun displayName(folder: Folder): String {
