+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/settings/account/FolderListPreference.kt
@@ -10,7 +10,7 @@ import androidx.core.content.res.TypedArrayUtils
 import androidx.preference.ListPreference
 import app.k9mail.legacy.ui.folder.FolderNameFormatter
 import com.fsck.k9.ui.R
-import net.thunderbird.feature.folder.api.RemoteFolder
+import net.thunderbird.feature.mail.folder.api.RemoteFolder
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.koin.core.parameter.parametersOf
